# JaegerExporter: Performance Improvements