--- a/lib/Directory.Build.props
+++ b/lib/Directory.Build.props
@@ -1,8 +1,0 @@
-<Project>
-  <Import Project="..\Directory.Build.props" Condition="Exists('..\Directory.Build.props')" />
-  <Import Project="$([MSBuild]::GetDirectoryNameOfFileAbove($(MSBuildProjectDirectory), 'OpenTelemetry.sln'))\build\Common.prod.props" />
-
-  <PropertyGroup>
-    <SkipAnalysis>True</SkipAnalysis>
-  </PropertyGroup>
-</Project>