--- a/lib/Thrift/Protocols/TAbstractBase.cs
+++ b/lib/Thrift/Protocols/TAbstractBase.cs
@@ -23,10 +23,6 @@ namespace Thrift.Protocols
     // ReSharper disable once InconsistentNaming
     public interface TAbstractBase
     {
-#if NETSTANDARD2_1
         ValueTask WriteAsync(TProtocol tProtocol, CancellationToken cancellationToken);
-#else
-        Task WriteAsync(TProtocol tProtocol, CancellationToken cancellationToken);
-#endif
     }
 }