--- a/lib/Thrift/Settings.Stylecop
+++ b/lib/Thrift/Settings.Stylecop
@@ -1,5 +1,0 @@
-<StyleCopSettings Version="4.3" >
-  <GlobalSettings>
-    <BooleanProperty Name="RulesEnabledByDefault">False</BooleanProperty>
-  </GlobalSettings>
-</StyleCopSettings>