--- a/lib/Thrift/TBaseClient.cs
+++ b/lib/Thrift/TBaseClient.cs
@@ -1,4 +1,4 @@
-// Licensed to the Apache Software Foundation(ASF) under one
+﻿// Licensed to the Apache Software Foundation(ASF) under one
 // or more contributor license agreements.See the NOTICE file
 // distributed with this work for additional information
 // regarding copyright ownership.The ASF licenses this file
@@ -51,12 +51,12 @@ namespace Thrift
             get { return ++_seqId; }
         }
 
-        public virtual async Task OpenTransportAsync()
+        public virtual async ValueTask OpenTransportAsync()
         {
             await OpenTransportAsync(CancellationToken.None).ConfigureAwait(false);
         }
 
-        public virtual async Task OpenTransportAsync(CancellationToken cancellationToken)
+        public virtual async ValueTask OpenTransportAsync(CancellationToken cancellationToken)
         {
             if (!_inputProtocol.Transport.IsOpen)
             {