--- a/lib/Thrift/TMultiplexedProcessor.cs
+++ b/lib/Thrift/TMultiplexedProcessor.cs
@@ -1,4 +1,4 @@
-// Licensed to the Apache Software Foundation(ASF) under one
+﻿// Licensed to the Apache Software Foundation(ASF) under one
 // or more contributor license agreements.See the NOTICE file
 // distributed with this work for additional information
 // regarding copyright ownership.The ASF licenses this file
@@ -129,7 +129,11 @@ namespace Thrift
                 _msgBegin = messageBegin;
             }
 
+#if NETSTANDARD2_1
+            public override async ValueTask<TMessage> ReadMessageBeginAsync(CancellationToken cancellationToken)
+#else
             public override async Task<TMessage> ReadMessageBeginAsync(CancellationToken cancellationToken)
+#endif
             {
                 if (cancellationToken.IsCancellationRequested)
                 {