--- a/lib/Thrift/Thrift.csproj
+++ b/lib/Thrift/Thrift.csproj
@@ -15,6 +15,7 @@
 
   <ItemGroup>
     <PackageReference Include="System.Buffers" Version="4.5.0" Condition="'$(TargetFramework)' == 'net46' or '$(TargetFramework)' == 'netstandard2.0'" />
+    <PackageReference Include="System.Threading.Tasks.Extensions" Version="4.5.3" Condition="'$(TargetFramework)' == 'net46' or '$(TargetFramework)' == 'netstandard2.0'" />
   </ItemGroup>
 
   <ItemGroup Condition="'$(TargetFramework)' == 'net46'">