--- a/lib/Thrift/Transports/Client/TBufferedClientTransport.cs
+++ b/lib/Thrift/Transports/Client/TBufferedClientTransport.cs
@@ -55,11 +55,7 @@ namespace Thrift.Transports.Client
 
         public override bool IsOpen => !_isDisposed && _transport.IsOpen;
 
-#if NETSTANDARD2_1
         public override async ValueTask OpenAsync(CancellationToken cancellationToken)
-#else
-        public override async Task OpenAsync(CancellationToken cancellationToken)
-#endif
         {
             CheckNotDisposed();
 
@@ -73,11 +69,7 @@ namespace Thrift.Transports.Client
             _transport.Close();
         }
 
-#if NETSTANDARD2_1
         public override async ValueTask<int> ReadAsync(byte[] buffer, int offset, int length, CancellationToken cancellationToken)
-#else
-        public override async Task<int> ReadAsync(byte[] buffer, int offset, int length, CancellationToken cancellationToken)
-#endif
         {
             //TODO: investigate how it should work correctly
             CheckNotDisposed();
@@ -122,11 +114,7 @@ namespace Thrift.Transports.Client
             return await ReadAsync(buffer, offset, length, cancellationToken).ConfigureAwait(false);
         }
 
-#if NETSTANDARD2_1
         public override async ValueTask WriteAsync(byte[] buffer, int offset, int length, CancellationToken cancellationToken)
-#else
-        public override async Task WriteAsync(byte[] buffer, int offset, int length, CancellationToken cancellationToken)
-#endif
         {
             CheckNotDisposed();
 
@@ -181,11 +169,7 @@ namespace Thrift.Transports.Client
             }
         }
 
-#if NETSTANDARD2_1
         public override async ValueTask FlushAsync(CancellationToken cancellationToken)
-#else
-        public override async Task FlushAsync(CancellationToken cancellationToken)
-#endif
         {
             CheckNotDisposed();