--- a/lib/Thrift/Transports/Client/TSocketClientTransport.cs
+++ b/lib/Thrift/Transports/Client/TSocketClientTransport.cs
@@ -82,11 +82,7 @@ namespace Thrift.Transports.Client
             }
         }
 
-#if NETSTANDARD2_1
         public override async ValueTask OpenAsync(CancellationToken cancellationToken)
-#else
-        public override async Task OpenAsync(CancellationToken cancellationToken)
-#endif
         {
             if (cancellationToken.IsCancellationRequested)
             {