… OpenTelemetry.sln
… benchmarks/Benchmarks.csproj
… benchmarks/Exporter/JaegerExporterBenchmarks.cs
… benchmarks/OpenTelemetrySdkBenchmarks.cs
… benchmarks/Program.cs
… lib/Directory.Build.targets
… lib/Thrift/Collections/TCollections.cs
… lib/Thrift/Collections/THashSet.cs
… lib/Thrift/ITAsyncProcessor.cs
… lib/Thrift/ITProcessorFactory.cs
… lib/Thrift/Protocols/Entities/TField.cs
… lib/Thrift/Protocols/Entities/TList.cs
… lib/Thrift/Protocols/Entities/TMap.cs
… lib/Thrift/Protocols/Entities/TMessage.cs
… lib/Thrift/Protocols/Entities/TMessageType.cs
… lib/Thrift/Protocols/Entities/TSet.cs
… lib/Thrift/Protocols/Entities/TStruct.cs
… lib/Thrift/Protocols/Entities/TType.cs
… lib/Thrift/Protocols/ITProtocolFactory.cs
… lib/Thrift/Protocols/TCompactProtocol.cs
… lib/Thrift/Protocols/TJSONProtocol.cs
… lib/Thrift/Protocols/TProtocolDecorator.cs
… lib/Thrift/Protocols/TProtocolException.cs
… lib/Thrift/Protocols/Utilities/TBase64Helper.cs
… lib/Thrift/Protocols/Utilities/TBase64Utils.cs
… lib/Thrift/Protocols/Utilities/TJsonProtocolConstants.cs
… lib/Thrift/Protocols/Utilities/TJsonProtocolHelper.cs
… lib/Thrift/Protocols/Utilities/TProtocolUtil.cs
… lib/Thrift/README.md
… lib/Thrift/Server/TBaseServer.cs