`JaegerUdpBatcher`, and the Thrift implementation it uses, is a total memory hog. When it is fed a span it allocates a buffer, writes the span into it, and then grabs the final size that will be needed off the buffer. It then throws away that buffer and puts the span into a list. During flush it goes through the list and re-writes the spans into another buffer. That buffer is then copied into yet another buffer on the UDP socket. Lots and lots of buffering going on. There's also over-use of Tasks. SendAsync on a socket really just writes into a buffer and returns synchronously.

So my theory going into this was...

1) We can reduce some work and memory usage by saving that initial buffer used to calculate size for the final transmission instead of saving the raw span and re-processing it.
2) We can reduce GC pressure by pooling buffers.
3) We can reduce allocations by using ValueTask where everything is actually running synchronously. Updated the ApacheThrift library for this part.

Benchmark:
```csharp
        [Benchmark]
        public async Task JaegerExporter_Batching()
        {
            using (var jaegerUdpBatcher = new JaegerUdpBatcher(
                new JaegerExporterOptions
                {
                    MaxPacketSize = int.MaxValue,
                    MaxFlushInterval = TimeSpan.FromHours(1)
                },
                new BlackHoleTransport()))
            {
                jaegerUdpBatcher.Process = new OpenTelemetry.Exporter.Jaeger.Implementation.Process("TestService", null);

                for (int i = 0; i < this.NumberOfBatches; i++)
                {
                    for (int c = 0; c < this.NumberOfSpans; c++)
                    {
                        await jaegerUdpBatcher.AppendAsync(this.testSpan, CancellationToken.None).ConfigureAwait(false);
                    }

                    await jaegerUdpBatcher.FlushAsync(CancellationToken.None).ConfigureAwait(false);
                }
            }
        }
```

Before:
``` ini

BenchmarkDotNet=v0.12.0, OS=Windows 10.0.18363
Intel Core i9-9880H CPU 2.30GHz, 1 CPU, 16 logical and 8 physical cores
.NET Core SDK=3.1.200-preview-014977
  [Host]     : .NET Core 3.1.1 (CoreCLR 4.700.19.60701, CoreFX 4.700.19.60801), X64 RyuJIT
  DefaultJob : .NET Core 3.1.1 (CoreCLR 4.700.19.60701, CoreFX 4.700.19.60801), X64 RyuJIT


```
|                  Method | NumberOfBatches | NumberOfSpans |        Mean |     Error |    StdDev |        Gen 0 |      Gen 1 |     Gen 2 |   Allocated | Completed Work Items | Lock Contentions |
|------------------------ |---------------- |-------------- |------------:|----------:|----------:|-------------:|-----------:|----------:|------------:|---------------------:|-----------------:|
| **JaegerExporter_Batching** |               **1** |         **10000** |    **352.7 ms** |   **6.72 ms** |   **6.28 ms** |   **24000.0000** |  **6000.0000** |         **-** |   **193.35 MB** |               **3.0000** |                **-** |
| **JaegerExporter_Batching** |              **10** |         **10000** |  **3,530.4 ms** |  **56.53 ms** |  **52.88 ms** |  **242000.0000** | **27000.0000** | **1000.0000** |  **1933.44 MB** |               **2.0000** |                **-** |
| **JaegerExporter_Batching** |             **100** |         **10000** | **35,705.5 ms** | **654.64 ms** | **612.35 ms** | **2423000.0000** | **94000.0000** | **7000.0000** | **19334.34 MB** |               **5.0000** |                **-** |

After:
``` ini

BenchmarkDotNet=v0.12.0, OS=Windows 10.0.18363
Intel Core i9-9880H CPU 2.30GHz, 1 CPU, 16 logical and 8 physical cores
.NET Core SDK=3.1.200-preview-014977
  [Host]     : .NET Core 3.1.1 (CoreCLR 4.700.19.60701, CoreFX 4.700.19.60801), X64 RyuJIT
  DefaultJob : .NET Core 3.1.1 (CoreCLR 4.700.19.60701, CoreFX 4.700.19.60801), X64 RyuJIT


```
|                  Method | NumberOfBatches | NumberOfSpans |        Mean |     Error |    StdDev |       Gen 0 |     Gen 1 | Gen 2 |  Allocated | Completed Work Items | Lock Contentions |
|------------------------ |---------------- |-------------- |------------:|----------:|----------:|------------:|----------:|------:|-----------:|---------------------:|-----------------:|
| **JaegerExporter_Batching** |               **1** |         **10000** |    **146.5 ms** |   **2.92 ms** |   **5.55 ms** |   **1000.0000** |         **-** |     **-** |   **10.94 MB** |               **1.0000** |                **-** |
| **JaegerExporter_Batching** |              **10** |         **10000** |  **1,421.6 ms** |  **19.13 ms** |  **16.95 ms** |  **13000.0000** |         **-** |     **-** |   **107.3 MB** |               **3.0000** |                **-** |
| **JaegerExporter_Batching** |             **100** |         **10000** | **14,385.2 ms** | **217.99 ms** | **203.91 ms** | **134000.0000** | **1000.0000** |     **-** | **1070.93 MB** |               **4.0000** |                **-** |

Basically it is faster and uses less memory.