@SergeyKanzhelev 2020-02-19
@CodeBlanch while you are on it - do you think it may be a good idea to build Thrift into the exporter code instead of a separate NuGet? I never published it on nuget.org. Mostly because doesn't want confusion.