@CodeBlanch 2020-02-19
@SergeyKanzhelev It definitely crossed my mind when I was in there updating all the files. There's a whole bunch of stuff in the Thirft library we're not using. None of the server stuff, none of the reading stuff, none of the transports (we have our own transport in the JaegerExporter). Could probably shed a lot of it if we pulled in just what is needed. We use the TCompactProtocol and the entities, basically. We're really just building a buffer for the batch and then dropping that on a UDP socket (buffer). A lot (all?) of the asynchronicity in Thirft is just overhead for us because we're not actually doing any IO.