@SergeyKanzhelev 2020-02-19
@CodeBlanch Thrift used here is basically a copy of https://github.com/jaegertracing/jaeger-client-csharp/tree/master/src/Thrift/netcore/Thrift. Since we are changing our version I'd suggest we just compile it into the exporter code. Maybe as a separate PR

@yurishkuro do you have more context and/or suggestions here?