@yurishkuro 2020-02-19
cc @Falco20019 re thrift/nuget question

Re the top-level performance concern - it sounds like the implementation is suboptimal. In some other Jaeger clients we usually use just two buffers, one for each span to get the encoded size, and the final buffer with the combined payload. By going a bit lower-level in Thrift encoding it's possible to use just a single buffer where the spans are appended in the final encoding.

Is it necessary to support Jaeger format directly? There's always the path of otel-sdk -> otel-collector -> jaeger.