@CodeBlanch 2020-02-20
Random question, IJaegerUpdBatcher has this interface:
```csharp
    public interface IJaegerUdpBatcher : IDisposable
    {
        Process Process { get; }

        Task<int> AppendAsync(JaegerSpan span, CancellationToken cancellationToken);

        Task<int> CloseAsync(CancellationToken cancellationToken);

        Task<int> FlushAsync(CancellationToken cancellationToken);
    }
```
CloseAsync just calls FlushAsync and isn't called ever. Should we drop it?