@Falco20019 2020-02-20
@CodeBlanch CloseAsync should be called in Dispose to ensure the spans are send to the agent even if not closed manually before. Otherwise they might get lost if not yet flushed. That’s at least how it‘s implemented in Jaeger. Haven’t looked at the implementation here so far.

@yurishkuro Have to check more into your question once in the office.