@CodeBlanch 2020-02-22
**Updates...**

1. Updated JaegerUdpBatcher to call CloseAsync during Dispose based on what @Falco20019 said. Tested, turns out Dispose is never called on JaegerUdpBatcher because BatchingSpanProcessor and SimpleSpanProcessor don't dispose of their Exporter when they are disposed. Fixed that up.

2. Switched to the official [ApacheThrift](https://www.nuget.org/packages/ApacheThrift) package. The performance looked good so I removed the fork. The package is a bit tricky because it bundles a net35 & net45 binary which has a different API. I had to do some kungfu in the csproj so we only use the netstandard2.0 binary.

**Work remaining...**

I have to merge in #486 which is going to be a bit of effort. I'll work on that this weekend.