@CodeBlanch 2020-02-22
Update...

* #486 merged into this PR.
* Switched ApacheThrift reference to PrivateAssets="All".
* Updated the description & performance numbers on the PR description to reflect latest master vs this branch.