@CodeBlanch 2020-02-27
@SergeyKanzhelev My goal is for the JaegerExporter to allocate enough memory for 1 batch and then re-use that for all subsequent batches it sends. Still messing with it, I'll open a new PR if I make significant headway.